Reorganized CUDA/OptiX backends into 'core' and 'api' part
This commit refactors the CUDA/OptiX backend code into parts that are
related to dynamic resolution of API symbols `{cuda,optix}_api.{h,cpp}`
and the remaining functionality (`{cuda,optix}_core.cpp`).

Some CUDA backend-specific bits were moved from `init.cpp` to
`cuda_core.cpp`.
